+++ tests/smoke/CSharp70/UseOutVariablesInMethodInvocations/OutInMethodsClass.cs
@@ -1,15 +1,19 @@
-﻿using System.CodeDom;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 
 namespace CSharp70.UseOutVariablesInMethodInvocations
 {
-    static class OutClass
+    static class OutInMethodsClass
     {
         public static bool Method(int i, out int j, ref int l)
         {
             j = 0;
             l = l == 0 ? 1: 0;
+            return true;
+        }
+
+        public static bool Method(int i)
+        {
             return true;
         }
 
@@ -30,6 +34,19 @@ namespace CSharp70.UseOutVariablesInMethodInvocations
         {
             i = 0;
             return Enumerable.Empty<T>();
+        }
+
+        public static int MethodInt(int i, out int j, ref int l)
+        {
+            j = 0;
+            l = l == 0 ? 1 : 0;
+            return 0;
+        }
+
+        public static int MethodInt(int i, out int j)
+        {
+            j = 0;
+            return 0;
         }
 
         public static bool Bool(bool value) => value;